Refactor commands.

+++ uos/commands/commands_filesystem.py
@@ -1,27 +1,12 @@
+from .basecommand import BaseCommand
 from ..uos import UOS
 
 
-class UserFilesystem:
-    def user_filesystem_commands(self):
-        return {
-            "CREATE ?": self.command_create_help,
-            'CREATE DIR': self.command_create_dir,
-            'CREATE FILE': self.command_create_file,
-            'CHANGE DIR': self.command_change_dir,
-            'DELETE DIR': self.command_delete_dir,
-            'DELETE FILE': self.command_delete_file,
-            'DELETE ?': self.command_delete_help,
-            'EDIT FILE': self.command_edit_file,
-            'EDIT ?': self.command_edit_help,
-            'MOUNT': self.command_mount,
-            'MOVE DIR': self.command_move_dir,
-            'MOVE FILE': self.command_move_file,
-            'RENAME DIR': self.command_rename_dir,
-            'RENAME FILE': self.command_rename_file,            
-            'RENAME ?': self.command_rename_help,
-        }
-
+class FilesystemCommands(BaseCommand):
     def command_create_dir(self, dirname, location=None):
+        if not self.clearance(1):
+            return
+
         newdir = UOS.drive.Path(dirname, location)
         if newdir.isdir():
             self.writer_add('Directory already exists')
@@ -30,6 +15,9 @@ class UserFilesystem:
             self.writer_add("{0} directory was created".format(dirname))
 
     def command_create_file(self, filename, filetype, location=None):
+        if not self.clearance(1):
+            return
+
         filepath = UOS.drive.Path((filename, filetype), location)
         if filepath.exists():
             line = "{0} file already exits.".format(filename)
@@ -38,6 +26,9 @@ class UserFilesystem:
             self.link.action.flip('Editor', filepath.path)
 
     def command_create_help(self):
+        if not self.clearance(1):
+            return
+
         self.writer_clear()
         self.writer_add( ["USAGE:",
                      "CREATE FILE",
@@ -53,6 +44,9 @@ class UserFilesystem:
                "CREATE ?"])
 
     def command_change_dir(self, *dirs):
+        if not self.clearance(1):
+            return
+
         source = UOS.drive.Path()
         if source.change_dir(*dirs):
             if source.isdir():
@@ -64,6 +58,9 @@ class UserFilesystem:
             self.writer_add('Does not exists')
 
     def command_delete_dir(self, dirname, location=None):
+        if not self.clearance(1):
+            return
+
         dirpath = UOS.drive.Path(dirname, location)
         if dirpath.isdir():
             dirpath.rmdir()
@@ -72,6 +69,9 @@ class UserFilesystem:
             self.writer_add('{0} directory does not exists'.format(dirname))
 
     def command_delete_file(self, filename, filetype, location=None):
+        if not self.clearance(1):
+            return
+
         filepath = UOS.drive.Path((filename, filetype), location)
         if filepath.isfile():
             filepath.remove()
@@ -80,6 +80,9 @@ class UserFilesystem:
             self.writer_add("File does't exists")
 
     def command_delete_help(self):
+        if not self.clearance(1):
+            return
+
         self.writer_clear()
         self.writer_add( ["USAGE:",
                      "DELETE FILE",
@@ -95,6 +98,9 @@ class UserFilesystem:
                      "DELETE ?"] )
 
     def command_edit_file(self, filename, filetype, location=None):
+        if not self.clearance(1):
+            return
+
         filepath = UOS.drive.Path((filename, filetype), location)
         if filepath.isfile():
             self.link.action.flip('Editor', filepath.path, True)
@@ -102,6 +108,9 @@ class UserFilesystem:
             self.writer_add('Unable to find {}.{}'.format(filename, filetype))
 
     def command_edit_help(self):
+        if not self.clearance(1):
+            return
+
         self.writer_clear()
         self.writer_add( ["USAGE:",
                           "EDIT FILE",
@@ -111,18 +120,24 @@ class UserFilesystem:
                           "EDIT ?"])
 
     def command_mount(self):
+        if not self.clearance(1):
+            return
+
         print('look for external storage')
 
     def command_move_dir(self, source):
+        if not self.clearance(1):
+            return
+
         filepath = UOS.drive.Path(source)
         if filepath.isdir():
             self.info.filepath = filepath
-            self.link.state = self.command_move_dir_new
+            self.link.state = self.move_dir_new
             self.writer_add("Enter new location")
         else:
             self.writer_add("Invalid directory")
 
-    def command_move_dir_new(self, dest):
+    def move_dir_new(self, dest):
         self.link.state = None
         destpath = UOS.drive.Path(self.info.filepath.basename(), dest)
         if destpath.exists():
@@ -132,15 +147,18 @@ class UserFilesystem:
             self.writer_add('Directory has been moved')
 
     def command_move_file(self, filename, filetype, location=None):
+        if not self.clearance(1):
+            return
+
         filepath = UOS.drive.Path((filename, filetype), location)
         if filepath.isfile():
             self.info.filepath = filepath
-            self.link.state = self.command_move_file_new
+            self.link.state = self.move_file_new
             self.writer_add("Enter new location")
         else:
             self.writer_add("Invalid directory")
 
-    def command_move_file_new(self, dest):
+    def move_file_new(self, dest):
         self.link.state = None
         filepath = UOS.drive.Path(self.info.filepath.basename(), dest)
         if filepath.exists():
@@ -150,15 +168,18 @@ class UserFilesystem:
             self.writer_add('File has been moved')
 
     def command_rename_dir(self, dirname, location=None):
+        if not self.clearance(1):
+            return
+
         dirpath = UOS.drive.Path(dirname, location)
         if dirpath.isdir():
-            self.link.state = self.command_rename_dir_new
+            self.link.state = self.rename_dir_new
             self.info.filepath = dirpath
             self.writer_add('Enter new directory name')
         else:
             self.writer_add("Directory doesn't exists")
 
-    def command_rename_dir_new(self, dirname, location=None):
+    def rename_dir_new(self, dirname, location=None):
         self.link.state = None
         dirpath = UOS.drive.Path(dirname, location)
         if not dirpath.exists():
@@ -168,17 +189,20 @@ class UserFilesystem:
             self.writer_add("Directory already exists")
 
     def command_rename_file(self, filename, filetype, location=None):
+        if not self.clearance(1):
+            return
+
         self.info.data = filetype
         self.info.name = location
         filepath = UOS.drive.Path((filename, filetype), location)
         if filepath.isfile():
             self.writer_add("Enter new filename for " + filename)
             self.info.filepath = filepath
-            self.link.state = self.command_rename_file_new
+            self.link.state = self.rename_file_new
         else:
             self.writer_add("{0} doesn't exists".format(filename))
 
-    def command_rename_file_new(self, filename, force=False):
+    def rename_file_new(self, filename, force=False):
         self.link.state = None
         filepath = UOS.drive.Path((filename, self.info.data), self.info.name)
         if filepath.exists() and not force:
@@ -188,6 +212,9 @@ class UserFilesystem:
             self.writer_add('{0} has been rename to {1}'.format(self.info.name, filename))
 
     def command_rename_help(self):
+        if not self.clearance(1):
+            return
+
         self.writer_clear()
         self.writer_add( ["USAGE:",
                      "RENAME FILE",
